+++ composer/trainer/trainer_hparams.py
@@ -10,10 +10,8 @@ import dataclasses
 import datetime
 import logging
 import os
-import sys
-import tempfile
 import warnings
-from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, List, Optional, Sequence, Tuple, Type, Union, cast
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, List, Optional, Sequence, Tuple, Union, cast
 
 import torch
 import yahp as hp
@@ -889,53 +887,3 @@ class ExperimentHparams(hp.Hparams):
         ]
 
         return trainer, fit_kwargs, eval_kwargs
-
-
-def _warning_on_one_line(message: str, category: Type[Warning], filename: str, lineno: int, file=None, line=None):
-    # From https://stackoverflow.com/questions/26430861/make-pythons-warnings-warn-not-mention-itself
-    return f'{category.__name__}: {message} (source: {filename}:{lineno})\n'
-
-
-def train_via_hparams() -> None:
-    """Entrypoint to train a model via hparams."""
-    warnings.formatwarning = _warning_on_one_line
-
-    if len(sys.argv) == 1:
-        sys.argv.append('--help')
-
-    hparams = TrainerHparams.create(cli_args=True)  # reads cli args from sys.argv
-
-    trainer = hparams.initialize_object()
-
-    # if using wandb, store the config inside the wandb run
-    try:
-        import wandb
-    except ImportError:
-        pass
-    else:
-        if wandb.run is not None:
-            wandb.config.update(hparams.to_dict())
-
-    # Only log the config once, since it should be the same on all ranks.
-    if dist.get_global_rank() == 0:
-        with tempfile.NamedTemporaryFile(mode='x+') as f:
-            f.write(hparams.to_yaml())
-            trainer.logger.file_artifact(
-                LogLevel.FIT,
-                artifact_name=f'{trainer.state.run_name}/hparams.yaml',
-                file_path=f.name,
-                overwrite=True,
-            )
-
-    # Print the config to the terminal and log to artifact store if on each local rank 0
-    if dist.get_local_rank() == 0:
-        print('*' * 30)
-        print('Config:')
-        print(hparams.to_yaml())
-        print('*' * 30)
-
-    trainer.fit()
-
-
-if __name__ == '__main__':
-    train_via_hparams()
